mips: retranslate when CP0 Config5 MSAEn bit changes
If MSA is enabled or disabled then any translated code including MSA
instructions needs to be retranslated so that the output of the
check_msa_access function is correct for the current CPU state.
Otherwise when an OS enables MSAEn and returns to a userland task which
is executing an MSA instruction, the existing translation will be
reused and retrigger a now invalid MSA Disabled exception.

The MIPS_HFLAG_BMASK_* fields were required moving in order to keep
MIPS_HFLAG_TMASK fields contiguous with the new MIPS_HFLAG_MSA field.
They were moved up 12 bits to the top of hflags to free up bits for any
further TMASK bits that may be required in the future without moving
them again.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>