[TCG ISSUE] Workaround for gen_muldiv instructions
Strange behaviour noticed in non-singlestep mode in madd, maddu, msub tests.
Different value is written into cpu_HI depending on place where temp is freed
(but in both cases temp is not used after freeing!).

Workaround: instructions which use cpu_HI will free temps at the end of
translation.

This seems to be already fixed in 1.7.0 QEMU qemu, but there were a lot of tcg
changes. Expecting this to be fixed once rebased onto head-of-tree QEMU.