target-mips: add MTHC0 and MFHC0 for MIPS32
The MIPS32r5 architecture adds two new instructions for accessing the upper
32-bits of the selected 64-bit CP0 register.

The MFHC0 instruction performs a move-from-high Coprocessor 0 operation which
moves the contents from the upper 32-bits of a coprocessor 0 register that has
been extended by 32-bits, to a generate purpose register.

The MTHC0 instruction performs a move-to-high Coprocessor 0 operation which
moves the contents of a general register to the upper 32-bits of a coprocessor 0
register that has been extended by 32-bits.

Signed-off-by: Leon Alrae <leon.alrae@imgtec.com>