mips: add support Global Interrupt Controller
This patch implements a part of the Global Interrupt Controller (GIC)
included in the Multi-core Coherent Processing System. The GIC handles
the distribution of interrupts between and among the CPUs in the Cluster.
On a CPU with multiple VPEs, each VPE has its own set of interrupt inputs.
The GIC has the ability to route interrupts to each VPE independently and
generally does not know or care to which CPU a VPE belongs.

I6400 core definition is updated for the GIC.

Updates #68

Signed-off-by: Yongbok Kim <yongbok.kim@imgtec.com>