target-mips: make ITC Configuration Tags accessible to the CPU
Add CP0.ErrCtl register with WST, SPR and ITC bits. In 34K and interAptiv
processors these bits are used to enable CACHE instruction access to
different arrays. When WST=0, SPR=0 and ITC=1 the CACHE instruction will
access ITC tag values.

Generally we do not support caches and we have been treating it as nop.
But since CACHE can operate on ITC Tags new MIPS_HFLAG_ITC_CACHE hflag
is introduced to generate the helper only when CACHE is in the ITC Access
mode.

Signed-off-by: Leon Alrae <leon.alrae@imgtec.com>