hw/mips: add ITC Storage Cells
Define single structure for FIFO and Semaphore cells. Read-only FIFO bit
in the ITC cell tag indicates the type of the cell. If the ITC Storage
contains both types of cells then FIFOs are located before Semaphores.

Since issuing thread can get blocked on the access to a cell (in E/F
Synchronized and P/V Synchronized Views) each cell contains a bitmap
to track which threads are blocked so that they will get woken up when
cell's state changes and clears the blocking condition.

Signed-off-by: Leon Alrae <leon.alrae@imgtec.com>