target-mips: Fix nanoMIPS addiu[32] erroneous sign extension
Further to the preceeding commit ad5ed727b53c ("R7 ADDIU32 takes an
unsigned 16-bit immediate not signed"), stop using gen_arith_imm which
takes the immediate as an int16_t parameter. This caused the unsigned
immediate to be cast back to a signed one which was then incorrectly
sign extended to 32 or 64 bits.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Fixes: ad5ed727b53c ("R7 ADDIU32 takes an unsigned 16-bit immediate not signed")