target-mips: Implement nanoMIPS EXTW instruction
Implement the nanoMIPS EXTW instruction, which is similar to the MIPS r6
ALIGN instruction except that it counts the other way and in bits
instead of bytes. We therefore generalise gen_align() into
gen_align_bits() (which counts in bits instead of bytes and optimises
when bits = size of the word), and implement gen_align() and a new
gen_ext() based on that. Since we need to know the word size to check
for when the number of bits == the word size, the opc argument is
replaced with a wordsz argument (either 32 or 64).

Signed-off-by: James Hogan <james.hogan@imgtec.com>
(cherry picked from commit 92e0432a76fb7bfdde64e6f53e2e0b7fa4cd9031)